requirements.txt
streamlit
openai
python-dotenv
pandas
PyPDF2